Ensure connections are always re-used
The current implementation of RedisHelper causes a new instance of
StrictRedis to be initialized every time get_atomic_connection() is
called. This results in a new connection to Redis being established for
each check against the ratelimits.

With these changes StrictRedis is only initialized once, resulting in
connection reuse because StrictRedis maintains an internal connection
pool.